+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -17,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feature - Added statistics visualizations graph for timetabling analysis
 - Feature - Added statistics visualization button for each satisfiable instance.
 - Feature - Added 3 new alternative solvers for the system.
+- Feature - Added SCIPSolver for alternative solution.
 
 ### Changed
 - Updated statistics visualization buttons and their information display options.